Refactoring. Added french localization made by sacaman1

+++ ITEIT87x/ITEIT87x.h
@@ -83,7 +83,7 @@ class IT87xMonitor : public SuperIOMonitor
     OSDeclareDefaultStructors(IT87xMonitor)
 	
 private:
-    UInt8                   voltageGain;
+    float                   voltageGain;
     
     bool                    has16bitFanCounter;
